Re-reading OVSDB data when agent heart beats fail
The patch set converts the Monitor L2gateway agent to the
transact agent when the Neutron server is disconnected
from the Rabbit MQ bus, and the L2gateway Monitor agent
is not able to report its state to the Neutron server.
When the Neutron server connects back to the bus, it selects
a particular agent as the Monitor agent and the OVSDB server
data is re-read so as to get the up-to-date view of the data.

Change-Id: Idd662d35ffee8e91cba7800b88d27359f3d98138