l2-gateway-connection-create reports OVSDBError
The patch set fixes
1. a problem described in the
bug 1440301 where the L2gateway agent throws
a syntax error as the logical switch is passed
as a list inside a list.
2. a problem where the database model
cannot store the value of the string
"unspecified-fault" in 16 character column width
for the port fault status and switch fault status.
3. a problem where the L2gateway agent and the plugin
do not handle empty port and switch fault status
which makes validation of the logical gateway fail
when an l2-gateway-connection is created.

Closes-Bug:1440301
Change-Id: Id4530b26407f2a6d57de7254476d77a36bb70a27