Fix for the launchpad bug 1440649
The patch set fixes the problem described
in the launchpad bug 1440649 where deletion
of an existing l2gw connection fails.
The code logic in the service plugin (inside
the method delete_port_mac) handled
the case where the incoming argument was a dict.
It, however, did not handle the case where
the incoming argument was a list.

Closes-Bug: 1440649
Change-Id: Id315416b12b6af581aa768c1378fae6496c342f5