Fix _get_l2_gateway call
If an invalid L2 gateway UUID is passed to _get_l2_gateway method,
it returns None instead of raising L2GatewayNotFound exception.
This leads to an Internal Server Error.

This patch replaces the 'try...except' block with an 'if' statement
because get method of SQLAlchemy Query object returns a None
instead of raising NotFound exception.

Closes-Bug: #1570146
Change-Id: Ib95fdaa4af613800159d89960b8b98c839508eaf