Tag the alembic migration revisions for Liberty
Previously when we had one repo with one alembic branch we would
create a milestone revision on that single branch. Now we have
multiple repos and expand/contract branches for each repo.
So from now on we tag the final revision on every branch when we make
a milestone release.

Change-Id: I6c76e56afd85f6d52ded663d1e812d3df91fed83
Closes-bug: #1571045