Tag the alembic migration revisions for Mitaka
Previously when we had one repo with one alembic branch we would
create a milestone revision on that single branch. Now we have
multiple repos and expand/contract branches for each repo.
So from now on we tag the final revision on every branch when we make
a milestone release.

Since no new migration scripts were added during the Mitaka cycle
in networking-l2gw, this patch adds the tag in Liberty expand
and contract scripts.

Change-Id: I720f1089dc725f8e3856725a6a4334115a44cd24
Closes-Bug: #1581158