Adding OVSLIB
This is the first of serious of commits aimed to add ovslib to L2GW.
The current commit lays the foundation for L2GW OVSDB manager class to use instead
of the current OVSDB writer/monitor.

simple unit test was added to test the code to make sure all the moving parts (IDL, Session, Connection, Transaction, etc.)
are connected properly.

Change-Id: I14107a0677030b5230adcfc856bfb386a6ab8592