Fix the configuration of the tempest plugin
Normalize the name of the section of the l2gw-specific options
(lowercase) and use explicitely the config wrapper from tempest.
Remove some code used for registering options, which is most likely
already handled internally in Tempest.
Without the fixes, some tests are failing while initializing the
configuration (CONF.L2GW) when the runner is non-serial.

Also, reorder the module imports to follow the Style Guideline.

Change-Id: I668ca20ef8a0371324173506427849603ebcdcc3