Support quotas
Ensure that get_l2_gateway_connections_count is added to support
quotas. A change in horizon/neutron calls this method and
if its not an admin tenant then this fails. This results in a
tenant being unable to create a FIP when using horizon.

Closes-bug: #1762414

Change-Id: I8875334fa6700c33154bc690ebcad6cb5fe6c0c2
(cherry picked from commit ff8ff91e25b71a3453039731fcee57397ca3be10)