Retire networking-l2gw: remove zuul jobs
networking-l2gw is not a neutron stadium project, and not governed by
openstack, so it is time to move it to x/ namespace. As a first step
disable zuul jobs to avoid collisions when the new repository will be
created as x/networking-l2gw.

Change-Id: I1bf048df49c385a92fea55f99bd0a7c4283960cf
(cherry picked from commit 454cf61f8a191fc25a1d85d7be5df81eab405c54)